requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.36
langchain-openai==0.1.22

openai==1.51.0

httpx==0.27.0

streamlit==1.37.1
faiss-cpu==1.8.0
tiktoken==0.7.0
python-dotenv==1.0.1

unstructured==0.14.9
requests==2.32.3
pydantic==2.7.4
PyYAML==6.0.1
